+++ packages/cli/src/cmd/run.ts
@@ -1,3 +1,4 @@
+import { ElementConfig } from '@flood/element-core/src/ElementOption'
 import { Argv, Arguments, CommandModule } from 'yargs'
 import { checkFile } from './common'
 import {
@@ -7,7 +8,6 @@ import {
 	ElementOptions,
 } from '@flood/element-core'
 import { runCommandLine as runMultipleUser } from '@flood/element-scheduler'
-
 import chalk from 'chalk'
 import { EventEmitter } from 'events'
 import { ReportCache } from '@flood/element-report'
@@ -19,19 +19,42 @@ import { resolve, dirname, basename, extname, join } from 'path'
 
 interface RunCommonArguments extends Arguments, ElementRunArguments {}
 
-async function getAllTestScriptsFromConfiguration(
-	args: RunCommonArguments,
-): Promise<RunCommonArguments> {
-	const fileErr = checkFile(args.configFile, 'Configuration file')
+async function getConfigurationFromConfig(args: RunCommonArguments): Promise<RunCommonArguments> {
+	const { file, configFile, _, $0, mu } = args
+	const fileErr = checkFile(configFile, 'Configuration file')
 	if (fileErr) throw fileErr
-	const { options, paths } = await readConfigFile(args.configFile)
+	const configFileFromArgs: ElementConfig = await readConfigFile(configFile)
+	const { options, paths, testSettings } = configFileFromArgs
+	let testFiles: string[]
+	let notExistingFiles: string[]
 
-	if (!paths.testPathMatch || !paths.testPathMatch.length) {
-		throw Error('Found no test scripts matching testPathMatch pattern')
+	if (file) {
+		testFiles = [file]
+		notExistingFiles = []
+	} else {
+		if (!paths.testPathMatch || !paths.testPathMatch.length) {
+			throw Error('Found no test scripts matching testPathMatch pattern')
+		}
+		const filesPattern = getFilesPattern(paths.testPathMatch)
+		testFiles = filesPattern.files.sort()
+		notExistingFiles = filesPattern.notExistingFiles.sort()
 	}
-	const { files, notExistingFiles } = getFilesPattern(paths.testPathMatch)
 
-	return { ...options, paths, testFiles: files.sort(), notExistingFiles: notExistingFiles.sort() }
+	return {
+		...options,
+		testSettings,
+		testFiles,
+		notExistingFiles,
+		runArgs: {
+			...args,
+		},
+		_,
+		$0,
+		file,
+		'fail-status-code': args['fail-status-code'],
+		configFile,
+		mu,
+	}
 }
 
 const cmd: CommandModule = {
@@ -44,7 +67,7 @@ const cmd: CommandModule = {
 			if (!file) {
 				console.log(
 					chalk.redBright(
-						`The mode 'running the test with a config file' does not support running with multiple users`,
+						`The mode 'running the test with a multiple files' does not support running with multiple users`,
 					),
 				)
 				process.exit(0)
@@ -55,7 +78,12 @@ const cmd: CommandModule = {
 			await runMultipleUser(opts)
 			process.exit(0)
 		}
-		const runArgs = file ? args : await getAllTestScriptsFromConfiguration(args)
+		const runArgs = await getConfigurationFromConfig(args)
+		if (runArgs.fastForward && runArgs.slowMo) {
+			console.error(chalk.redBright(`Arguments fast-forward and slow-mo are mutually exclusive`))
+			process.exit(0)
+		}
+
 		const result = await runSingleUser(runArgs)
 
 		if (args.export) {