+++ packages/core/src/ElementOption.ts
@@ -13,6 +13,23 @@ import { PlaywrightClient } from './driver/Playwright'
 import { BrowserType } from './page/types'
 import ms from 'ms'
 
+export interface RunArguments {
+	loopCount?: number
+	duration?: string | number
+	stepDelay?: string | number
+	actionDelay?: string | number
+	fastForward?: boolean
+	slowMo?: boolean
+	watch?: boolean
+	file?: string
+	headless?: boolean
+	devtools?: boolean
+	sandbox?: boolean
+	verbose?: boolean
+	browser?: BrowserType
+	export?: boolean
+}
+
 export interface ElementRunArguments {
 	testFiles: string[]
 	file: string
@@ -22,6 +39,7 @@ export interface ElementRunArguments {
 	devtools?: boolean
 	sandbox?: boolean
 	loopCount?: number
+	duration?: string | number
 	stepDelay?: string | number
 	actionDelay?: string | number
 	fastForward?: boolean
@@ -36,6 +54,9 @@ export interface ElementRunArguments {
 	export?: boolean
 	showScreenshot?: boolean
 	notExistingFiles: string[]
+	mu: boolean
+	runArgs: RunArguments
+	testSettings?: TestSettings
 }
 
 export interface ElementOptions {
@@ -49,6 +70,7 @@ export interface ElementOptions {
 	sandbox: boolean
 	process?: NodeJS.Process
 	verbose: boolean
+	testSettings: TestSettings
 	testSettingOverrides: TestSettings
 	testObserverFactory?: (t: TestObserver) => TestObserver
 	persistentRunner: boolean
@@ -57,6 +79,23 @@ export interface ElementOptions {
 	browser: BrowserType
 	export?: boolean
 	showScreenshot?: boolean
+}
+
+export interface ElementConfig {
+	options: ElementOptions
+	paths: {
+		workRoot: string
+		testDataRoot: string
+		testPathMatch: string[]
+	}
+	flood?: {
+		hosted: boolean
+		vu: number
+		duration: number
+		rampup: number
+		regions: string[]
+	}
+	testSettings?: TestSettings
 }
 
 function getWorkRootPath(file: string, root?: string): string {
@@ -82,7 +121,8 @@ function setupDelayOverrides(
 	testSettingOverrides: TestSettings,
 ): TestSettings {
 	if (testSettingOverrides == null) testSettingOverrides = {}
-	const { actionDelay, stepDelay } = args
+	const actionDelay = args.runArgs?.actionDelay ?? args.actionDelay
+	const stepDelay = args.runArgs?.stepDelay ?? args.stepDelay
 
 	if (typeof actionDelay === 'string' && actionDelay) {
 		testSettingOverrides.actionDelay = ms(actionDelay)
@@ -96,10 +136,10 @@ function setupDelayOverrides(
 		testSettingOverrides.stepDelay = stepDelay
 	}
 
-	if (args.fastForward) {
+	if (args.runArgs?.fastForward || args.fastForward) {
 		testSettingOverrides.stepDelay = 1000
 		testSettingOverrides.actionDelay = 1000
-	} else if (args.slowMo) {
+	} else if (args.runArgs?.slowMo || args.slowMo) {
 		testSettingOverrides.stepDelay = 10000
 		testSettingOverrides.actionDelay = 10000
 	}
@@ -148,10 +188,12 @@ export function normalizeElementOptions(
 	args: ElementRunArguments,
 	cache: ReportCache,
 ): ElementOptions {
-	const { file, verbose } = args
+	const { file, verbose, runArgs, testSettings } = args
 	const workRootPath = getWorkRootPath(file, args['work-root'])
 	const testDataPath = getTestDataPath(file, args['test-data-root'])
-	const verboseBool = !!verbose
+	let verboseBool = !!verbose
+
+	if (runArgs?.file && !!runArgs.verbose) verboseBool = !!runArgs.verbose
 
 	const reporter = verboseBool ? new VerboseReporter(cache) : new BaseReporter(cache)
 
@@ -173,12 +215,38 @@ export function normalizeElementOptions(
 		showScreenshot: args.showScreenshot,
 	}
 
+	opts.testSettings = { ...testSettings }
+
 	if (args.loopCount) {
-		opts.testSettingOverrides.loopCount = args.loopCount
+		opts.testSettings.loopCount = args.loopCount
 	}
+
+	if (args.duration) {
+		opts.testSettings.duration = args.duration
+	}
+
+	if (args.stepDelay) {
+		opts.testSettings.stepDelay = args.stepDelay
+	}
+
+	if (args.actionDelay) {
+		opts.testSettings.actionDelay = args.actionDelay
+	}
+
+	if (runArgs?.file) {
+		const { loopCount, duration } = runArgs
+		if (loopCount !== undefined) opts.testSettingOverrides.loopCount = loopCount
+		if (duration !== undefined) opts.testSettingOverrides.duration = duration
+		opts.headless = runArgs.headless ?? opts.headless
+		opts.devtools = runArgs.devtools ?? opts.devtools
+		opts.sandbox = runArgs.sandbox ?? opts.sandbox
+		opts.browser = runArgs.browser ?? opts.browser
+		opts.export = !!runArgs.export ?? opts.export
+	}
+
 	opts.testSettingOverrides = setupDelayOverrides(args, opts.testSettingOverrides)
 
-	if (args.watch) {
+	if ((runArgs?.file && runArgs?.watch) || args.watch) {
 		opts.persistentRunner = true
 		opts.testCommander = makeTestCommander(file)
 	}